requirements.txt
asgiref==3.4.1
dataclasses==0.8
Django==3.2.12
django-cors-headers==3.10.1
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.0
pymongo==4.0.2
python-decouple==3.6
pytz==2021.3
sqlparse==0.2.4
typing_extensions==4.1.1
